Rename concise_casts test to saturating_casts to better document what it does. (It still covers the concise casts as an add-on, but this does very little now that saturating casts are covered directly.)
Ensure that saturating_cast brings infinities into range for
floating-point types.

Make clamp produce an error if bounds are out of range for target type.

Update saturating casts test to handle remaining floating-pont cases.
